feat: add modules task
